+++ src/egui_app/ui/waveform_view/interactions/scrollbar.rs
@@ -8,7 +8,7 @@ pub(in super::super) fn render_waveform_scrollbar(
     ui: &mut Ui,
     scroll_rect: egui::Rect,
     view: WaveformView,
-    view_width: f32,
+    view_width: f64,
 ) {
     let palette = style::palette();
     let scroll_resp = ui.interact(
@@ -18,8 +18,8 @@ pub(in super::super) fn render_waveform_scrollbar(
     );
     let scroll_bg = style::with_alpha(palette.bg_primary, 140);
     ui.painter().rect_filled(scroll_rect, 0.0, scroll_bg);
-    let indicator_width = scroll_rect.width() * view_width;
-    let indicator_x = scroll_rect.left() + scroll_rect.width() * view.start;
+    let indicator_width = (scroll_rect.width() as f64 * view_width) as f32;
+    let indicator_x = scroll_rect.left() + (scroll_rect.width() as f64 * view.start) as f32;
     let indicator_rect = egui::Rect::from_min_size(
         egui::pos2(indicator_x, scroll_rect.top()),
         egui::vec2(indicator_width.max(8.0), scroll_rect.height()),
@@ -31,6 +31,6 @@ pub(in super::super) fn render_waveform_scrollbar(
         && let Some(pos) = scroll_resp.interact_pointer_pos()
     {
         let frac = ((pos.x - scroll_rect.left()) / scroll_rect.width()).clamp(0.0, 1.0);
-        app.controller.scroll_waveform_view(frac);
+        app.controller.scroll_waveform_view(frac as f64);
     }
 }
